+++ index.d.ts
@@ -1,15 +1,13 @@
+import {Options, ErrorObject} from "ajv"
 export default class Validator {
     static supportedVersions: Set<string>;
-    constructor(ajvOptions?: {
-        strict: boolean;
-        validateFormats: boolean;
-    });
+    constructor(ajvOptions?: Options);
     resolveRefs(opts?: {
         specification?: object
     }): object;
     validate(schema: object | string): Promise<{
         valid: boolean;
-        errors?: any
+        errors?: ErrorObject[];
     }>;
     specification: object;
     version: string;